split tests for default vs custom templates, add filtering tests
